+++ src/test/java/de/codescape/jira/plugins/scrumpoker/rest/entities/CardRepresentationTest.java
@@ -1,10 +1,10 @@
-package de.codescape.jira.plugins.scrumpoker.rest;
+package de.codescape.jira.plugins.scrumpoker.rest.entities;
 
 import org.junit.Test;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertThat;
 
 public class CardRepresentationTest {